pyproject.toml
[project]
name = "chat"
version = "0.1.0"
description = "chat"
authors = [{name = "Frans Fürst", email = "frans.fuerst+om-office@protonmail.com"}]
readme = "Readme.md"
requires-python = ">=3.13,<4"
dependencies = [
    "httpx>=0.28.1",
    "textual>=6.11.0",
]

[project.urls]
Repository = "https://github.com/frans-fuerst/ai-chat-tui.git"

[project.scripts]
chat = 'src.chat:main'

# see https://docs.astral.sh/uv/concepts/projects/dependencies/
[dependency-groups]
dev = [
    "ruff",
    "mypy",
    "ty",
    "pre-commit",
    "pytest",
    "types-pyyaml",
    "yamllint",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
markers = [
    "unit: Unit tests (deselect with '-m \"not unit\"')",
    "system: Tests involving an attached device (deselect with '-m \"not system\"')",
]
testpaths = ["src", "test"]

[tool.mypy]
python_version = "3.13"
strict=true
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
show_error_codes = true
exclude = '^(\\.venv)'
mypy_path="typeshed"
cache_dir = ".mypy_cache"
explicit_package_bases = true

[tool.ruff]
target-version = "py313"
line-length = 80

[tool.ruff.lint]
# see https://docs.astral.sh/ruff/linter/
# https://github.com/astral-sh/ruff-pre-commit
select = [
    "ALL",  # all available linting checks
    "B",    # flake8-bugbear: reports potential bugs
    "C90",  # Complexity: violations of 90th percentile of cyclomatic complexity
    "C",    # Complexity: general code complexity issues
    "E",    # pycodestyle: check PEP 8 style guidelines
    "F",    # pyflakes: detects potential errors like undefined/unused symbols
    "I",    # isort: check for proper import sorting
    "PL",   # pylint: Reports issues related to Pylint checks.
    "SIM",  # flake8-simplify: Suggests simplifications in code (e.g., replacing complex expressions with simpler ones).
    "UP",   # pyupgrade: Checks and suggests improvements to modernize Python code (e.g., f-strings).
    "W",    # Warning: General warnings or potentially problematic code patterns.
]
ignore = [
    # These should be fixed eventually
    "E501",     # line too long
    "ERA001",   # commented out code
    "FBT001",   # Boolean-typed positional argument in function definition
    "FBT002",   # Boolean default positional argument in function definition
    "PLR0913",  # Too many arguments in function definition
    "PLR2004",  # Magic value used in comparison, consider replacing `127` with a constant variable
    "D100",     # Missing docstring in public module
    "D102",     # Missing docstring in public method
    "D103",     # Missing docstring in public function
    "RUF012",   # Mutable class attributes should be annotated with `typing.ClassVar`

    # These are ignored on purpose
    "COM812",   # missing-trailing-comma
    "D203",     # incorrect-blank-line-before-class
    # "D211",   # no-blank-line-before-class
    "D205",     # 1 blank line required between summary line and description
    "D212",     # Multi-line docstring summary should start at the first line
    "D213",     # Multi-line docstring
    "D400",     # First line should end with a period
    "D401",     # First line of docstring should be in imperative mood: "Adds generator created from @player_definition_fn to new track definition"
    "D415",     # First line should end with a period, question mark, or exclamation point
    "EM101",    # Exception must not use a string literal
    "EM102",    # Exception must not use an f-string literal, assign to variable first
    "EXE001",   # Shebang is present but file is not executable
    "S101",     # Use of `assert` detected
    "SIM117",   # Use single `with`
    "TRY003",   # Avoid specifying long messages outside the exception class
]
